+++ js/main.js
@@ -48,9 +48,9 @@ define( [
     'SCENERY/nodes/VBox',
     
     'SCENERY/util/AccessibilityPeer',
-    'SCENERY/util/BasicNodeEvents',
     'SCENERY/util/CanvasContextWrapper',
     'SCENERY/util/Color',
+    'SCENERY/util/FixedNodeEvents',
     'SCENERY/util/Font',
     'SCENERY/util/LinearGradient',
     'SCENERY/util/Instance',
@@ -59,7 +59,6 @@ define( [
     'SCENERY/util/RenderInterval',
     'SCENERY/util/SceneImage',
     'SCENERY/util/SceneryStyle',
-    'SCENERY/util/SplitNodeEvents',
     'SCENERY/util/Trail',
     'SCENERY/util/TrailPointer',
     'SCENERY/util/Util',